// madnessfeed-client/src/app/types.ts
export interface MfaLoginOptions {
    totp: boolean
    passkey: boolean
    passkeyChallenge?: string
    passkeyCredentialIds: string[]
}

export interface PasskeyInfo {
    id: number
    name: string
    created: number
    lastUsed?: number
}

export interface MfaStatus {
    totpEnabled: boolean
    passkeys: PasskeyInfo[]
}

export interface TotpSetupResponse {
    secret: string
    uri: string
}

export interface PasskeyRegistrationOptions {
    challenge: string
    rpName: string
    userId: string
    userName: string
    algorithms: number[]
    excludeCredentialIds: string[]
}

export interface PasskeyRegistrationRequest {
    name: string
    clientDataJSON: string
    attestationObject: string
}